+++ example/simple.dart
@@ -1,4 +1,5 @@
 import 'dart:io';
+import 'dart:math';
 
 import 'package:dart_dice_parser/dart_dice_parser.dart';
 import 'package:logging/logging.dart';
@@ -13,7 +14,7 @@ void main() {
   });
 
   // create a roller for D20 advantage (roll 2d20, drop lowest)
-  final dice = DiceExpression.create('2d20-L');
+  final dice = DiceExpression.create('2d20-L', Random.secure());
 
   // each roll returns different results.
   final int result1 = dice.roll();