+++ lib/dart_dice_parser.dart
@@ -2,6 +2,7 @@
 ///
 ///
 /// Usage example:
+///
 /// ```dart
 /// const input = '2d20-L'; // D20 advantage -- roll 2d20, drop lowest
 /// final diceExpr = DiceExpression.create(input);
@@ -13,3 +14,5 @@
 library dart_dice_parser;
 
 export 'src/dice_expression.dart';
+export 'src/dice_roller.dart';
+export 'src/results.dart';