+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# ?.?.?
+- add compounding dice: `5d6!!`
+- add keep high/low: `2d20k`
+- remove `!!` as 'limited explosion'
+- fix syntax for exploding dice (2d6!, not 2d!6). 
+- remove exploding method from dice roller -- now it's part of the AST. 
+- dice rolls return RollResult
+- clean redundant parser config
+
 # 4.1.0
 - cleanup examples
 - add syntax for `>=` and `<=` for counts, drop, clamp